Rename BaseInputFeedbackCard to BaseFeedbackCard
