+++ wls-gui-wahllokalsystem/src/components/common/cards/BaseFeedbackCard.vue
