+++ wls-gui-wahllokalsystem/src/components/wahlhandlung/TheUnguetilgeWahlscheineVerifyCard.vue
@@ -15,7 +15,7 @@
           @update:model-value="onWahlscheinnummerChanged"
         />
 
-        <base-input-feedback-card
+        <base-feedback-card
           v-if="feedbackWahlscheinIsGueltigIsVisible"
           title="Wahlschein ist gültig"
           :type="InputFeedbackTypeEnum.success"
@@ -34,9 +34,9 @@
               Wählerverzeichnis.
             </li>
           </ul>
-        </base-input-feedback-card>
+        </base-feedback-card>
 
-        <base-input-feedback-card
+        <base-feedback-card
           v-if="feedbackWahlscheinIsUngueltigIsVisible"
           :title="titleFeedbackWahlscheinUngueltig"
           :type="InputFeedbackTypeEnum.error"
@@ -82,7 +82,7 @@
               <v-spacer />
             </v-row>
           </template>
-        </base-input-feedback-card>
+        </base-feedback-card>
         <v-img
           position="left"
           style="height: 40mm"
@@ -90,19 +90,19 @@
           :src="wahlscheinExampleImage"
         />
 
-        <base-input-feedback-card
+        <base-feedback-card
           v-if="feedbackLoadingFailedIsVisible"
           title="Liste ungültiger Wahlscheine nicht verfügbar"
           :type="InputFeedbackTypeEnum.error"
           >Die ungültigen Wahlscheine konnten nicht geladen
-          werden.</base-input-feedback-card
+          werden.</base-feedback-card
         >
-        <base-input-feedback-card
+        <base-feedback-card
           v-if="feedbackNoDataAvailableIsVisible"
           title="Liste ungültiger Wahlscheine nicht verfügbar"
           :type="InputFeedbackTypeEnum.error"
           >Der Inhalt der Datei mit den ungültigen Wahlscheinen ist
-          leer.</base-input-feedback-card
+          leer.</base-feedback-card
         >
       </v-form>
     </v-card-text>
@@ -134,7 +134,7 @@ import wahlscheinExampleImage from "@/assets/previewWahlscheinnummerOnWahlschein
 import BaseButtonRefresh from "@/components/common/buttons/BaseButtonRefresh.vue";
 import BaseButtonSave from "@/components/common/buttons/BaseButtonSave.vue";
 import BaseTextButton from "@/components/common/buttons/BaseTextButton.vue";
-import BaseInputFeedbackCard from "@/components/common/cards/BaseInputFeedbackCard.vue";
+import BaseFeedbackCard from "@/components/common/cards/BaseFeedbackCard.vue";
 import BaseNumberInput from "@/components/common/inputs/BaseNumberInput.vue";
 import { useRules } from "@/composables/common/rules.ts";
 import { MAX_LENGTH_FOR_TEXT_INPUT } from "@/constants.ts";
